Issue #1440 - update get_block in database_api and condenser_api
