+++ libraries/plugins/apis/block_api/include/steemit/plugins/block_api/block_api_args.hpp
@@ -9,8 +9,19 @@
 
 namespace steemit { namespace plugins { namespace block_api {
 
-/* get_block */
+/* get_block_header */
 
+struct get_block_header_args
+{
+   uint32_t block_num;
+};
+
+struct get_block_header_return
+{
+   optional< block_header > header;
+};
+
+/* get_block */
 struct get_block_args
 {
    uint32_t block_num;
@@ -22,6 +33,12 @@ struct get_block_return
 };
 
 } } } // steemit::block_api
+
+FC_REFLECT( steemit::plugins::block_api::get_block_header_args,
+   (block_num) )
+
+FC_REFLECT( steemit::plugins::block_api::get_block_header_return,
+   (header) )
 
 FC_REFLECT( steemit::plugins::block_api::get_block_args,
    (block_num) )
