+++ libraries/plugins/apis/condenser_api/CMakeLists.txt
@@ -8,6 +8,7 @@ target_link_libraries( condenser_api_plugin
    account_by_key_api_plugin
    account_history_api_plugin
    database_api_plugin
+   block_api_plugin
    follow_api_plugin
    market_history_api_plugin
    network_broadcast_api_plugin
