+++ libraries/plugins/apis/condenser_api/include/steemit/plugins/condenser_api/condenser_api_plugin.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <steemit/plugins/json_rpc/json_rpc_plugin.hpp>
 #include <steemit/plugins/database_api/database_api_plugin.hpp>
+#include <steemit/plugins/block_api/block_api_plugin.hpp>
 
 #define STEEM_CONDENSER_API_PLUGIN_NAME "condenser_api"
 
@@ -11,7 +12,7 @@ using namespace appbase;
 class condenser_api_plugin : public appbase::plugin< condenser_api_plugin >
 {
 public:
-   APPBASE_PLUGIN_REQUIRES( (steemit::plugins::json_rpc::json_rpc_plugin)(steemit::plugins::database_api::database_api_plugin) )
+   APPBASE_PLUGIN_REQUIRES( (steemit::plugins::json_rpc::json_rpc_plugin)(steemit::plugins::database_api::database_api_plugin)(steemit::plugins::block_api::block_api_plugin) )
 
    condenser_api_plugin();
    virtual ~condenser_api_plugin();
