+++ libraries/plugins/apis/database_api/include/steemit/plugins/database_api/database_api.hpp
@@ -18,24 +18,6 @@ class database_api
       ~database_api();
 
       DECLARE_API(
-         /////////////////////////////
-         // Blocks and transactions //
-         /////////////////////////////
-
-         /**
-         * @brief Retrieve a block header
-         * @param block_num Height of the block whose header should be returned
-         * @return header of the referenced block, or null if no matching block was found
-         */
-         (get_block_header)
-
-         /**
-         * @brief Retrieve a full, signed block
-         * @param block_num Height of the block to be returned
-         * @return the referenced block, or null if no matching block was found
-         */
-         (get_block)
-
 
          /////////////
          // Globals //
